+++ lib/x509asn1.c
@@ -518,7 +518,7 @@ static const char *GTime2str(const char *beg, const char *end)
                        beg, beg + 4, beg + 6,
                        beg + 8, beg + 10, sec1, sec2,
                        fracl? ".": "", fracl, fracp,
-                       sep, tzl, tzp);
+                       sep, (int)tzl, tzp);
 }
 
 /*
@@ -558,7 +558,7 @@ static const char *UTime2str(const char *beg, const char *end)
   return curl_maprintf("%u%.2s-%.2s-%.2s %.2s:%.2s:%.2s %.*s",
                        20 - (*beg >= '5'), beg, beg + 2, beg + 4,
                        beg + 6, beg + 8, sec,
-                       tzl, tzp);
+                       (int)tzl, tzp);
 }
 
 /*
